MSN: Skylogin UNICODE fixes.
git-svn-id: http://svn.miranda-ng.org/main/trunk@16519 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/MSN/src/skylogin/common.h
@@ -82,7 +82,7 @@ typedef  struct
 typedef struct
 {
 	uchar	*Memory;
-	int 	MsZ;
+	unsigned long MsZ;
 }	Memory_U;
 
 typedef struct
